Broker View bug fixed and web interface improvement
Previously there was a bug that if a broker X had a number of topics
and after a reassignment it didn't have any topic anymore, it would
keep appearing to have the same topics (partitions replicas) as before
because the map brokertopicPartitions wasn't updated for such brokers
in the method updateView():Unit in BrokerViewActor.scala. It has now
been fixed by performing the right mapping for such brokers.