Operating System Metrics included in Broker Metrics
Long story short, after allowing the user to manually set partitions
assignment among the brokers, the need to better analyze the brokers
arise because one of the motivations to manually assign partitions is
to understand some of the broker's low level stats such as CPU load
and memory available, and thus improve the distributions of partitions
among the brokers by taking into consideration such informations.
That being said, kafka's MBeans are not the only ones being used now;
java.lang OperatingSystem MBean is used to get Operating System metrics
through JMX.

* New class OSMetric has been created to keep the Operating System
stats.
* Class BrokerMetrics now also wraps a OSMetric object.
* New KafkaMetrics method getOSMetric(MBeanServerConnection) was
created to fetch OS stats through JMX.
* New View app/views/common/expandedBrokerMetrics.scala.html was
created, based on brokerMetrics.scala.html, to display OS stats and to
only display the Mean of other metrics, the 1, 5, or 15 minutes
threshold are not displayed in this view.

TODO: Implement a filtering, possibly with AngularJS to better filter the
metrics and topics seen, and decrease the visual clutter.