Visual improvement and filtering of topics and metrics
Two filters were added to Manual Partitions Assignment: one to filter
topics, and the other to filter metrics thus reducing the visual
clutter of the page.

A new button was also added: Suggest Partition Assignment, which must
be implemented yet. When clicking on it, a suggestion of configuration
taking into consideration the metrics available must be set on the form.